Fix glass material (?)
Definitely it resembles more to expected output but the if statement
I used here is unintuitive. I am checking depth > 0, that it uses shadow rays
for non-primary rays. I was expecting that it'd only use it for primary rays
however for primary rays the outcome is brighter. Perhaps the mistake is
somewhere else that I should apply depth == 0 once I spot the problem