Fix glass material (*)
THIS TIME FOR REAL. It took me a single swap of code lines. The problem
was: I was starting directly with ambient light in get_color( ) instead of
a zero vector. That means in every recursion, ambient light was added
on radiance. However we want to add ambient light only once! I thought
it is weird to not include ambient( ) in shade_diffuse( ) because it was
added in get_color( ) so I wanted to make it more clear, accidentally
solving the bug I've inserted way before.