Create Scene composite
In my 5th or 6th trial, finally decoupling deserialized json from additional
scene data is possible. I used data composition pattern (not sure if that's
the official name) to create a composite Scene struct, that holds deserialized
data (trick here, it holds a reference, so there are lifetime annotations 'a,
hopefully this avoids copying the whole data because serde beautifully
deserializes with no copy, and I just don't want to mess it up). Now the
intention here is to add BVH and other related data under Scene struct
that is not part of json deserialization.