Update BVH tree node types to HeapBBoxable
So I have a problem, how do we actually track which shape's bounding box
we hit? Intersect just returns the boolean and I'm not sure where to update
hit_record. Since I want to use BBoxable trait, a possible solution is making
node type BBoxable (but we cannot directly do that because of the
unknown size at compilation time, wrap it around a pointer Arc)