Remove indices( ) from Shape
This was a method to access indices data of the shapes but that's not
really a good practice in Rust as far as I observe. Because now I want to
impl Shape for BBox but bounding box has no indices data. Frankly
this method was not called in anywhere so I just removed it