Refactor intersection and fix sphere hitrecords
I guess the same fix will be needed for Triangles -- wait no they are
called by Meshes, and it is correct. OK now ellipsoids and spheres
appear as they should. that ray_t parameter issue we had in Meshes
was prevalent in sphere intersection as well, so rays were not actually
reaching their world position (I mean it makes sense, t is the parameter
for rays, it needs to be correct). -- Oh wait there is still an issue with
dielectric. I sense it's because of hitrecord's entry point. Stay tuned.