store hitrecords in a pool
I needed to use mutex for the first time in my life! because
the parallelized render loop now requires a mutable
parameter, so for thread safety mutex was needed.
for a simple scene like lobster there are >1M hitrecords
I'm not sure if this post-processing experiment will be
meaningful at all.