Fix json parser error in absence of point lights
cornellbox_area.json was throwing error because no point light was
found. Currently it is fixed by deriving serde(default) however this
is not ideal because it does not tell if there is a new field in json (i.e.
area light is not handled by code, it just silently runs without warning
about unexpected field. which could be great to be able to render
without implementing all new modifications but also we could easily
forget about new fields if they were not given in the homework, so it
is not ideal for future use)