add motion blur
Yay I got it finally! Literally it is almost one line of code
but figuring out where exactly I should do the transform
took me a while. This was also because the current
hit( ) calls are spaghetti, I was going to transform the
ray at TLAS, before getting inside BLAS but that
requires some mutabilitiy changes I dreaded to
implement.