relocate decalibration
in order to introduce normal and bumpmapping i needed to copy the logic
I used in shade_diffuse, instead of copying I've relocated that part
and changed function signatures to adapt the change. now
hopefully we can have an exhaustive match arms for decal_mode
instead of separating it for brdf-related and other mappings.